requirements.txt
marshmallow>=3
importlib-metadata==4.13.0
flask==2.2.5
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
apispec[yaml]
apispec-webframeworks
tox
celery[redis]>=5.0.0
gunicorn
cloudinary
geopy

